AnToanLaoDongBE/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# MongoDB
motor>=3.3.0
beanie>=1.26.0

# OpenAI
openai>=1.40.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
aiofiles>=24.1.0

# PDF Report Generation
reportlab>=4.0.0

# HTTP Client
httpx>=0.27.0

# Kling AI (JWT auth for video generation)
PyJWT>=2.8.0

# Utilities
python-dateutil>=2.9.0

# Auth
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# QR codes for certificates
qrcode[pil]>=7.4.2
